Fix: Settings byte size

+++ src/Backend/Shared/Settings.cs
@@ -127,10 +127,10 @@ public static class Settings
     public static string CsvOutputPath { get; set; } = "";
 
     [ConfigKey("MAX_CONCURRENT_CONNECTIONS")]
-    public static ushort MaxConcurrentConnections { get; set; }
+    public static uint MaxConcurrentConnections { get; set; }
 
     [ConfigKey("MAX_CONSUMER_FETCH")]
-    public static ushort ConsumerFetchMax { get; set; }
+    public static uint ConsumerFetchMax { get; set; }
 
     [ConfigKey("MAX_PRODUCER_LINECOUNT")]
     public static ulong ProducerLineMax { get; set; }
